Extract logic for creating an integer attribute
