+++ lib/kitchen/terraform/config_attribute/command_timeout.rb
@@ -14,10 +14,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-require "kitchen/terraform/config_attribute"
-require "kitchen/terraform/config_attribute_cacher"
-require "kitchen/terraform/config_attribute_definer"
-require "kitchen/terraform/config_schemas/integer"
+require "kitchen/terraform/config_attribute_type/integer"
 
 # This attribute controls the number of seconds that the plugin will wait for Terraform commands to finish running.
 #
@@ -25,31 +22,9 @@ require "kitchen/terraform/config_schemas/integer"
 # Required:: False
 # Default:: +600+
 # Example:: <code>command_timeout: 1200</code>
-#
-# @abstract It must be included by a plugin class in order to be used.
-module ::Kitchen::Terraform::ConfigAttribute::CommandTimeout
-  # A callback to define the configuration attribute which is invoked when this module is included in a plugin class.
-  #
-  # @param plugin_class [::Kitchen::Configurable] A plugin class.
-  # @return [void]
-  def self.included(plugin_class)
-    ::Kitchen::Terraform::ConfigAttributeDefiner
-      .new(
-        attribute: self,
-        schema: ::Kitchen::Terraform::ConfigSchemas::Integer
-      )
-      .define plugin_class: plugin_class
-  end
-
-  # @return [::Symbol] the symbol corresponding to the attribute.
-  def self.to_sym
-    :command_timeout
-  end
-
-  extend ::Kitchen::Terraform::ConfigAttributeCacher
-
-  # @return [::Integer] 600.
-  def config_command_timeout_default_value
-    600
-  end
-end
+::Kitchen::Terraform::ConfigAttribute::CommandTimeout =
+  ::Kitchen::Terraform::ConfigAttributeType::Integer
+    .create(
+      attribute: :command_timeout,
+      default_value: 600
+    )
