+++ lib/kitchen/terraform/config_attribute_type.rb
@@ -14,18 +14,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-require "kitchen/terraform/config_attribute_type/integer"
+require "kitchen/terraform"
 
-# This attribute controls the number of concurrent operations to use while Terraform
-# {https://www.terraform.io/docs/internals/graph.html#walking-the-graph walks the resource graph}.
-#
-# Type:: {http://www.yaml.org/spec/1.2/spec.html#id2803828 Integer}
-# Required:: False
-# Default:: +10+
-# Example:: <code>parallelism: 50</code>
-::Kitchen::Terraform::ConfigAttribute::Parallelism =
-  ::Kitchen::Terraform::ConfigAttributeType::Integer
-    .create(
-      attribute: :parallelism,
-      default_value: 10
-    )
+# The namespace for configuration attribute types.
+module ::Kitchen::Terraform::ConfigAttributeType
+end
